xiao ascension refactor
